requirements.txt
asgiref==3.8.1
attrs==24.2.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
dj-rest-auth==7.0.0
Django==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.28.0
drf-standardized-errors==0.14.1
google-auth==2.36.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
python-decouple==3.8
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
rsa==4.9
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
